Update example to use "passthrough gltfs"
This showcases the ability to specify a body geometry with a gltf file.
The gltf file can articulate elaborate materials (multiple materials per
mesh, PBR materials, etc.) and that data is preserved through registration
and made available by the RenderEngineGltfClient.

To that end:

 - Introduce the example .gltf file (with accompanying sdf and package
   files).
 - Update the model directives to use them.

Most critically, this requires changes the server.
 - Where we rotate the scene after loading (to account for an apparent VTK
   bug), we explicitly specify the rotation pivot to be less sensitive to
   how the gltf is articulated.
 - The passthrough gltf may have an arbitrary hierarchy. We can't rely
   on objects being direct children of the "Renderer Node". Instead, we
   explicitly decorate all mesh nodes with a custom property:
   "from_drake":True. The test in label settings relies on this.

TODO:
 - Unit tests on server.py?